backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
behave==1.2.6
blinker==1.9.0
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudscraper==1.2.71
colorama==0.4.6
cryptography==46.0.5
cssselect==1.4.0
distro==1.9.0
exceptiongroup==1.3.1
execnet==2.1.2
fasteners==0.20
filelock==3.25.0
Flask==3.1.3
flask-cors==6.0.2
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.66.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.71.2
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mycdp==1.3.2
nest-asyncio==1.6.0
outcome==1.3.0.post0
packaging==26.0
parameterized==0.9.0
parse==1.21.1
parse_type==0.6.6
pdbp==1.8.2
platformdirs==4.9.4
playwright==1.58.0
pluggy==1.6.0
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
pynose==1.5.5
pyotp==2.9.0
pyparsing==3.3.2
PySocks==1.7.1
pytest==9.0.2
pytest-html==4.0.2
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==16.1
pytest-xdist==3.8.0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
resend==2.23.0
rich==14.3.3
sbvirtualdisplay==1.4.0
schedule==1.2.2
selenium==4.41.0
seleniumbase==4.47.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
tabcompleter==1.4.0
tenacity==9.1.4
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
wsproto==1.3.2
